refactor: CLI 아키텍처 분리 — __init__.py 2842줄 → repl.py + tool_handlers.py 추출
Clean Architecture: CLI 레이어에서 anthropic SDK 직접 참조 제거.
LLM 예외 타입을 core.llm.client에서 re-export하여 포트 계층 유지.

- __init__.py (2842줄) → 1488줄로 축소
- repl.py (423줄): _interactive_loop, _read_multiline_input, _build_prompt_session,
  _restore_terminal, signal handling 추출
- tool_handlers.py (957줄): _build_tool_handlers + 38개 핸들러 함수 추출
- result_cache.py (83줄): ResultCache 클래스 추출
- agentic_loop.py: import anthropic → LLMTimeoutError 등 래퍼 사용
- nl_router.py: import anthropic → LLMAuthenticationError 등 래퍼 사용
- core/llm/client.py: LLM 예외 타입 6종 re-export 추가
